Remove ancestry column Add parent_id column

+++ app/models/db_tree_view.rb
@@ -1,8 +1,8 @@
 class DbTreeView < ApplicationRecord
   include Stateable
   include Destroyable
+  include Treeable
   include TreeViewValidations
-  has_ancestry
 
   def self.apply_changes(data)
     errors = []
@@ -10,6 +10,7 @@ class DbTreeView < ApplicationRecord
       data.each do |_index, node_data|
         new_node = find_or_initialize_by(id: node_data[:id])
         new_node.assign_attributes(node_data.except(:id))
+        next if new_node.parent.disabled? && !new_node.persisted?
         if new_node.disabled?
           new_node.destroy
         end
